primitives/ed25519: Allow small order R by default
On second thought, this package might as well just match the "Algorithm
2" definition from the "Taming the many EdDSAs" paper.

While this is different from the default behavior of the old oasislabs
package, that particular bridge was burnt when I also chose to reject
non-canonical A as mandated by some of the specifications, and
interoperability is possible via `VerifyOptions` anyway.

This has no impact on the security of the scheme.