primitives/ed25519: Change ed25519consensus-like batch api
This is not compatible due to the following differences:

 * `NewBatchVerifier` returns a pointer.
 * `BatchVerifier.Verify` returns information about which signature(s)
   in the batch were invalid on failure, and supports specifying an
   entropy source.

Added calls:

 * `BatchVerifier.VerifyBatchOnly` equivalent to the ed25519consensus
   `BatchVerifier.Verify` call, with provisions for specifying an
   entropy source.
 * `BatchVerifier.AddWithOptions` for Ed25519ph/Ed25519ctx support.

While it is possible to just use `VerifyBatchOnly`, and manually verify
signatures in the event of a failure (since most callers will presumably
want to know which signature failed), doing so results in R, A, S and
H(R,A,m) being decompressed/deserialized/derived twice.  This
implementation opts for a slightly heavier `entry` struct to be able to
reuse the values derived as part of the batch verification process.