curve: Switch back to packing hardcoded tables
Yes, I've gone back and forth on this.  The difference from the last
time I did this is that it is stored in packed form and converted to
unpacked form on module initialization, rather than unpacking at lookup
time.

Unpacking the various tables takes sub-10 usec on my laptop, and this
reduces the maintenance burden.