primitives/ed25519: Pull in more tests from upstream
Upstream now explicitly tests that the verification behavior for edge
cases is consistent with ref10.  Use the same test vectors to exercise
all of the various profiles.

Note: The standard library downloads the test cases at test time by
fork/execing `go`.  That sucks so this just includes a copy.