curve: Add comments regarding non-canonical points
This situation is kind of a mess, but the pacakge needs to be able to
accept non-canonical points, since there are flavors of Ed25519
signatures that do not care about such things.

The current behavior should roughly be in line with what other libraries
do as well, and if people need to explicitly reject non-canonical
encodings, there are easy ways to do so (either use IsCanonical, or
check if the point round-trips).