primitives/ed25519: Optimize for very large batches
This makes several changes to the batch verification to improve
behavior when verifying large batches.

Internal changes:
 * Always (also) store -A as an Edwards point.
 * When creating a batch, if the batch is known to be over the size
   where precomputation will accelerate the multiscalar multiply
   operation, do not expand public keys in the entry.

This results in a negligible increase to memory consumption for batches
where expansion is applicable (due to the added overhead of storing -A),
and a dramatic decrease in memory consumption for large batches.

Additionally, there is a slight performance improvement for large
batches for the non-expanded case.

Note: `BatchVerfier.ForceNoPublicKeyExpansion` can be used to
force-disable the public key expansion.  This will be a net gain if
the batch is large and built via `Add`/`AddWithOptions`, or if
`VerifyBatchOnly` is used.